+++ masterSolution/PolyWars.API/Strategies/IMoveStrategy.cs
@@ -2,6 +2,6 @@
 
 namespace PolyWars.API.Strategies {
     public interface IMoveStrategy {
-        void Move(IMoveable moveable, decimal deltaTime);
+        void Move(IMoveable moveable, double deltaTime);
     }
 }
